requirements.txt
Django~=3.2.5
djangorestframework

psycopg2

# django-neomodel
# neo4j-driver
# neomodel